+++ CHANGES.md
@@ -5,3 +5,4 @@
 - Added: Core and Console projects
 - Added: Main class
 - Added: Query system / QueryHandler
+- Added: Getting work directory with path argument(0) if given